STM32 USBDevice: refactor files
