+++ features/unsupported/USBDevice/targets/TARGET_STM/USBHAL_IP_OTGFSHS.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2016 mbed.org, MIT License
+/* Copyright (c) 2017 mbed.org, MIT License
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy of this software
 * and associated documentation files (the "Software"), to deal in the Software without
@@ -15,13 +15,21 @@
 * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
-#ifndef USBHAL_STM32F769NI_H
-#define USBHAL_STM32F769NI_H
+#ifndef USBHAL_IP_OTGFSHS_H
+#define USBHAL_IP_OTGFSHS_H
+
+#if defined(TARGET_DISCO_F769NI)
 #define USBHAL_IRQn  OTG_HS_IRQn
-/*  must be multiple of 4 bytes */
-#define NB_ENDPOINT 4
+#else
+#define USBHAL_IRQn  OTG_FS_IRQn
+#endif
+
+#define NB_ENDPOINT  4 // Must be a multiple of 4 bytes
+
 #define MAXTRANSFER_SIZE  0x200
-#define FIFO_USB_RAM_SIZE (MAXTRANSFER_SIZE+MAX_PACKET_SIZE_EP0+MAX_PACKET_SIZE_EP1+MAX_PACKET_SIZE_EP2+MAX_PACKET_SIZE_EP3)
+
+#define FIFO_USB_RAM_SIZE (MAXTRANSFER_SIZE + MAX_PACKET_SIZE_EP0 + MAX_PACKET_SIZE_EP1 + MAX_PACKET_SIZE_EP2 + MAX_PACKET_SIZE_EP3)
+
 #if (FIFO_USB_RAM_SIZE > 0x500)
 #error "FIFO dimensioning incorrect"
 #endif
@@ -29,30 +37,37 @@
 typedef struct
 {
     USBHAL *inst;
+
     void (USBHAL::*bus_reset)(void);
     void (USBHAL::*sof)(int frame);
-    void (USBHAL::*connect_change)(unsigned int  connected);
+    void (USBHAL::*connect_change)(unsigned int connected);
     void (USBHAL::*suspend_change)(unsigned int suspended);
     void (USBHAL::*ep0_setup)(void);
     void (USBHAL::*ep0_in)(void);
     void (USBHAL::*ep0_out)(void);
     void (USBHAL::*ep0_read)(void);
     bool (USBHAL::*ep_realise)(uint8_t endpoint, uint32_t maxPacket, uint32_t flags);
-    bool (USBHAL::*epCallback[2*NB_ENDPOINT-2])(void);
-    /*  memorize dummy buffer used for reception */
-    uint32_t pBufRx[MAXTRANSFER_SIZE>>2];
-    uint32_t pBufRx0[MAX_PACKET_SIZE_EP0>>2];
-    uint8_t epComplete[2*NB_ENDPOINT];
-}USBHAL_Private_t;
+    bool (USBHAL::*epCallback[(2 * NB_ENDPOINT) - 2])(void);
+
+    uint8_t epComplete[2 * NB_ENDPOINT];
+
+    /* Memorize dummy buffer used for reception */
+    uint32_t pBufRx[MAXTRANSFER_SIZE >> 2];
+    uint32_t pBufRx0[MAX_PACKET_SIZE_EP0 >> 2];
+} USBHAL_Private_t;
 
 uint32_t HAL_PCDEx_GetTxFiFo(PCD_HandleTypeDef *hpcd, uint8_t fifo)
 {
     uint32_t len;
-    if (fifo == 0) len = hpcd->Instance->DIEPTXF0_HNPTXFSIZ>>16;
-    else
+    if (fifo == 0) {
+        len = hpcd->Instance->DIEPTXF0_HNPTXFSIZ >> 16;
+    }
+    else {
         len =  hpcd->Instance->DIEPTXF[fifo - 1] >> 16;
-    return len*4;
+    }
+    return len * 4;
 }
+
 void HAL_PCD_SOFCallback(PCD_HandleTypeDef *hpcd) 
 {
     USBHAL_Private_t *priv=((USBHAL_Private_t *)(hpcd->pData));
@@ -60,27 +75,31 @@ void HAL_PCD_SOFCallback(PCD_HandleTypeDef *hpcd)
     USB_OTG_GlobalTypeDef *USBx = hpcd->Instance;
     uint32_t sofnum = (USBx_DEVICE->DSTS & USB_OTG_DSTS_FNSOF) >> 8;
     void (USBHAL::*func)(int frame) = priv->sof;
-    /* fix me  call with same frame number */
     (obj->*func)(sofnum);
 }
 
-
-USBHAL * USBHAL::instance;
+USBHAL *USBHAL::instance;
 
 USBHAL::USBHAL(void) {
-    /*  init parameter  */
     USBHAL_Private_t *HALPriv = new(USBHAL_Private_t);
-    hpcd.Instance = USB_OTG_HS;
+
     memset(&hpcd.Init, 0, sizeof(hpcd.Init));
-    hpcd.Init.dev_endpoints = NB_ENDPOINT;
-    hpcd.Init.ep0_mps =   MAX_PACKET_SIZE_EP0;
+
+#if defined(TARGET_DISCO_F769NI)
+    hpcd.Instance = USB_OTG_HS;
     hpcd.Init.phy_itface = PCD_PHY_ULPI;
     hpcd.Init.Sof_enable = 0;
-
     hpcd.Init.speed = PCD_SPEED_HIGH;
-    //hpcd.Init.vbus_sensing_enable = 0;
-    //hpcd.Init.lpm_enable = 0;
-    /*  pass instance for usage inside call back */
+#else
+    hpcd.Instance = USB_OTG_FS;
+    hpcd.Init.phy_itface = PCD_PHY_EMBEDDED;
+    hpcd.Init.Sof_enable = 1;
+    hpcd.Init.speed = PCD_SPEED_FULL;
+#endif
+    hpcd.Init.dev_endpoints = NB_ENDPOINT;
+    hpcd.Init.ep0_mps = MAX_PACKET_SIZE_EP0;
+
+    // Pass instance for usage inside call back
     HALPriv->inst = this;
     HALPriv->bus_reset = &USBHAL::busReset;
     HALPriv->suspend_change = &USBHAL::suspendStateChanged;
@@ -91,7 +110,6 @@ USBHAL::USBHAL(void) {
     HALPriv->ep0_in = &USBHAL::EP0in;
     HALPriv->ep0_out = &USBHAL::EP0out;
     HALPriv->ep0_read = &USBHAL::EP0read;
-    hpcd.pData = (void*)HALPriv;
     HALPriv->epCallback[0] = &USBHAL::EP1_OUT_callback;
     HALPriv->epCallback[1] = &USBHAL::EP1_IN_callback;
     HALPriv->epCallback[2] = &USBHAL::EP2_OUT_callback;
@@ -99,16 +117,43 @@ USBHAL::USBHAL(void) {
     HALPriv->epCallback[4] = &USBHAL::EP3_OUT_callback;
     HALPriv->epCallback[5] = &USBHAL::EP3_IN_callback;
     instance = this;
-    /* Enable power and clocking */
+
+    // Configure USB pins and other clocks
+
+    // NUCLEO_144 boards
+#if defined(TARGET_NUCLEO_F207ZG) || \
+    defined(TARGET_NUCLEO_F412ZG) || \
+    defined(TARGET_NUCLEO_F429ZI) || \
+    defined(TARGET_NUCLEO_F446ZE) || \
+    defined(TARGET_NUCLEO_F767ZI) || \
+    defined(TARGET_NUCLEO_F746ZG)
+    __HAL_RCC_GPIOA_CLK_ENABLE();
+    pin_function(PA_8, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_FS));  /* OTG_FS_SOF */
+    pin_function(PA_9, STM_PIN_DATA(STM_MODE_INPUT, GPIO_NOPULL, GPIO_AF10_OTG_FS));  /* OTG_FS_VBUS */
+    pin_function(PA_10, STM_PIN_DATA(STM_MODE_AF_OD, GPIO_PULLUP, GPIO_AF10_OTG_FS)); /* OTG_FS_ID */
+    pin_function(PA_11, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_FS)); /* OTG_FS_DM */
+    pin_function(PA_12, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_FS)); /* OTG_FS_DP */
+    __HAL_RCC_USB_OTG_FS_CLK_ENABLE();
+
+#elif defined(TARGET_DISCO_L475VG_IOT01A) || \
+      defined(TARGET_DISCO_L476VG)
+    __HAL_RCC_GPIOA_CLK_ENABLE();
+    pin_function(PA_11, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_FS)); /* OTG_FS_DM */
+    pin_function(PA_12, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_FS)); /* OTG_FS_DP */
+    __HAL_RCC_GPIOC_CLK_ENABLE();
+    pin_function(PC_11, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_FS)); /* VBUS pin */
+    __HAL_RCC_PWR_CLK_ENABLE();
+    HAL_PWREx_EnableVddUSB();
+    __HAL_RCC_USB_OTG_FS_CLK_ENABLE();
+
+#elif defined(TARGET_DISCO_F769NI)
     __HAL_RCC_GPIOA_CLK_ENABLE();
     __HAL_RCC_GPIOB_CLK_ENABLE();
     __HAL_RCC_GPIOC_CLK_ENABLE();
     __HAL_RCC_GPIOH_CLK_ENABLE();
     __HAL_RCC_GPIOI_CLK_ENABLE();
-
     pin_function(PA_5, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // CLK
     pin_function(PA_3, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // D0
-
     pin_function(PB_0, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // D1
     pin_function(PB_1, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // D2
     pin_function(PB_5, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // D3
@@ -116,32 +161,41 @@ USBHAL::USBHAL(void) {
     pin_function(PB_11, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // D5
     pin_function(PB_12, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // D6
     pin_function(PB_13, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // D7
-
     pin_function(PC_0, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // STP
     pin_function(PH_4, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // NXT
     pin_function(PI_11, STM_PIN_DATA(STM_MODE_AF_PP, GPIO_NOPULL, GPIO_AF10_OTG_HS)); // DIR
-
     __HAL_RCC_USB_OTG_HS_ULPI_CLK_ENABLE();
     __HAL_RCC_USB_OTG_HS_CLK_ENABLE();
-
-    __HAL_RCC_SYSCFG_CLK_ENABLE();
-    hpcd.State = HAL_PCD_STATE_RESET;
-    HAL_PCD_Init(&hpcd);
-    /* 1.25kbytes */
-    /* min value 16 (= 16 x 4 bytes) */
-    /*  max value 256 (= 1K bytes ) */
-    /*  maximum sum is 0x140 */
-    HAL_PCDEx_SetRxFiFo(&hpcd, (MAXTRANSFER_SIZE/4));
-    /*  bulk/int 64 bytes in FS  */
-    HAL_PCDEx_SetTxFiFo(&hpcd, 0, (MAX_PACKET_SIZE_EP0/4)+1);
-    /*  bulk/int bytes in FS */
-    HAL_PCDEx_SetTxFiFo(&hpcd, 1, (MAX_PACKET_SIZE_EP1/4)+1);
-    HAL_PCDEx_SetTxFiFo(&hpcd, 2, (MAX_PACKET_SIZE_EP2/4));
-    /* ISOchronous */
-    HAL_PCDEx_SetTxFiFo(&hpcd, 3, (MAX_PACKET_SIZE_EP3/4));
-    NVIC_SetVector(USBHAL_IRQn, (uint32_t)&_usbisr);
-    NVIC_SetPriority(USBHAL_IRQn, 1);
-    HAL_PCD_Start(&hpcd);
-}
+#else
+#error "USB pins are not configured !"
 #endif
 
+    __HAL_RCC_SYSCFG_CLK_ENABLE();
+
+    // Configure PCD and FIFOs
+    hpcd.pData = (void*)HALPriv;
+    hpcd.State = HAL_PCD_STATE_RESET;
+    HAL_PCD_Init(&hpcd);
+
+    /* 1.25 kbytes */
+    /* min value 16 (= 16 x 4 bytes) */
+    /* max value 256 (= 1K bytes ) */
+    /* maximum sum is 0x140 */
+    HAL_PCDEx_SetRxFiFo(&hpcd, (MAXTRANSFER_SIZE / 4));
+    // EP0 = Bulk/Int 64 bytes in FS
+    HAL_PCDEx_SetTxFiFo(&hpcd, 0, (MAX_PACKET_SIZE_EP0 / 4) + 1);
+    // EP1 = Bulk/Int bytes in FS
+    HAL_PCDEx_SetTxFiFo(&hpcd, 1, (MAX_PACKET_SIZE_EP1 / 4) + 1);
+    // EP2
+    HAL_PCDEx_SetTxFiFo(&hpcd, 2, (MAX_PACKET_SIZE_EP2 / 4));
+    // EP3 = ISOchronous
+    HAL_PCDEx_SetTxFiFo(&hpcd, 3, (MAX_PACKET_SIZE_EP3 / 4));
+
+    // Configure interrupt vector
+    NVIC_SetVector(USBHAL_IRQn, (uint32_t)&_usbisr);
+    NVIC_SetPriority(USBHAL_IRQn, 1);
+
+    HAL_PCD_Start(&hpcd);
+}
+
+#endif // USBHAL_IP_OTGFSHS_H
